tidy: Wrapped document at 72 columns for ease of diffing
I realise that this is a pretty contentious practice these days, and
that many people strongly dislike it.  I've done it here because it
makes for much easier side-by-side diffs for the edits I'll be making.
Once these are done, if wrapping like this isn't acceptable, the
document can just be unwrapped.